pyproject.toml
[project]
name = "LLM_Survey_Prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polars",
    "marimo",
    "torch",
    "transformers",
    "accelerate",
    "safetensors",
    "huggingface_hub",
    "llama-cpp-python",
]
